chore(plot): rename timeseries_source to timeseries
